+++ next-docs/pages/how-to-contribute.tsx
@@ -12,17 +12,34 @@ export default function PageHowToContribute() {
         <h1 className="text-5xl font-semibold">How to contribute</h1>
         <h2 className="text-3xl font-semibold mt-8">Setup</h2>
 
+        <Text>Install `pnpm`:</Text>
+        <CodeSnippet>{`npm install -g pnpm`}</CodeSnippet>
+        <Text>
+          Check{' '}
+          <a className="underline" href="https://pnpm.io/installation">
+            pnpm docs
+          </a>{' '}
+          for additional ways.
+        </Text>
+
         <Text>Clone the repository:</Text>
         <CodeSnippet>{`git clone git@github.com:coingaming/moon-design.git`}</CodeSnippet>
 
         <Text>Install dependencies and link local packages together:</Text>
-        <CodeSnippet>{`yarn`}</CodeSnippet>
+        <CodeSnippet>{`pnpm install`}</CodeSnippet>
 
-        <Text>Compiles lib folder for each package:</Text>
-        <CodeSnippet>{`yarn build`}</CodeSnippet>
+        <Text>Compile and build all packages:</Text>
+        <CodeSnippet>{`pnpm run build`}</CodeSnippet>
 
-        <Text>Generates a static build of the style guide:</Text>
-        <CodeSnippet>{`yarn next dec`}</CodeSnippet>
+        <Text>Run storybook for development with Hot Module Replacement:</Text>
+
+        <CodeSnippet>{`pnpm run next dev`}</CodeSnippet>
+        <Text>For Hot Module Replacement run in parallel in separate tab:</Text>
+
+        <CodeSnippet>{`pnpm run watch`}</CodeSnippet>
+        <Text>Build and run storybook for production at localhost:80:</Text>
+        <CodeSnippet>{`pnpm run next build`}</CodeSnippet>
+        <CodeSnippet>{`pnpm run next start`}</CodeSnippet>
       </section>
       <section className="mt-8">
         <h2 className="text-3xl font-semibold mt-8">Check our githib</h2>